requirements.txt
flask==3.0.3
yt-dlp==2024.10.22
ffmpeg-python==0.2.0
flask-cors==5.0.0